unify field types react<->angular
