+++ packages/react/src/index.ts
@@ -28,6 +28,12 @@ export {
   ComponentFields,
   ComponentParams,
   EditMode,
+  ImageField,
+  ImageFieldValue,
+  ImageSizeParameters,
+  LinkField,
+  LinkFieldValue,
+  RichTextField,
 } from '@sitecore-content-sdk/content/layout';
 export { DictionaryPhrases, DictionaryService } from '@sitecore-content-sdk/content/i18n';
 export {
@@ -51,15 +57,9 @@ export {
   AppPlaceholderProps,
   renderEmptyPlaceholder,
 } from './components/Placeholder';
-export {
-  Image,
-  ImageProps,
-  ImageField,
-  ImageFieldValue,
-  ImageSizeParameters,
-} from './components/Image';
-export { RichText, RichTextProps, RichTextField } from './components/RichText';
-export { Text, TextField } from './components/Text';
+export { Image, ImageProps } from './components/Image';
+export { RichText, RichTextProps } from './components/RichText';
+export { Text } from './components/Text';
 export { DateField, DateFieldProps } from './components/Date';
 export {
   FEaaSComponent,
@@ -80,7 +80,7 @@ export {
 } from './components/FEaaS';
 export { DesignLibrary, DesignLibraryApp } from './components/DesignLibrary';
 export {} from './components/FEaaS/BYOCComponent';
-export { Link, LinkField, LinkFieldValue, LinkProps } from './components/Link';
+export { Link, LinkProps } from './components/Link';
 export { File, FileField } from './components/File';
 export {
   SitecoreProvider,
